Finally fixed the damn AI
Removed a set speed to 0 when not moving that was causing AI to never move.
Added conditions in Anim BP instead to cancel moving anim when standing still both in Boxxy and in BigArm.
Added a test character with a simple BT and BB for debugging purposes.
Regen navmesh based on change to capsules.
Fixed a compilation error with an interface event that somehow never got compiled.